Tests/LibWeb/Text/input/css/grid-row-start-tree-counting-function.html

<!doctype html>
<style>
    #foo {
        grid-row-start: sibling-count();
    }

    #bar {
        grid-row-start: span sibling-count();
    }
</style>
<script src="../include.js"></script>
<script>
    test(() => {
        println(document.styleSheets[0].cssRules[0].cssText);
        println(document.styleSheets[0].cssRules[1].cssText);
    });
</script>
