Tests/LibWeb/Text/expected/css/grid-row-start-tree-counting-function.txt

#foo { grid-row-start: sibling-count(); }
#bar { grid-row-start: span sibling-count(); }
